refactor(ward): get list api
