+++ src/hospital/hospital.service.ts
@@ -10,15 +10,12 @@ import { CreateHospitalDto } from './dto/request-dto/create-hospital.dto';
 import { Hospital } from './entities/hospital.entity';
 import { hash } from 'bcrypt';
 import { IdCheckDto } from './dto/request-dto/id-check.dto';
-import { CreateWardDto } from '../ward/dto/request-dto/create-ward.dto';
 import { Ward } from '../ward/entities/ward.entity';
 import { CreateRoomDto } from './dto/request-dto/create-room.dto';
 import { Room } from './entities/room.entity';
 import { Patient } from './entities/patient.entity';
 import { CreatePatientDto } from './dto/request-dto/create-patient.dto';
 import { Reservation } from '../reservation/entities/reservation.entity';
-import { UpdateWardDto } from '../ward/dto/request-dto/update-ward.dto';
-import { DeleteWardDto } from '../ward/dto/request-dto/delete-ward.dto';
 import { UpdateRoomDto } from './dto/request-dto/update-room.dto';
 import { DeleteRoomDto } from './dto/request-dto/delete-room.dto';
 import { UpdatePatientDto } from './dto/request-dto/update-patient.dto';
